Completed react-context clean-up-code(optional)
