+++ react-context/my-app/src/components/CartContext.tsx
@@ -1,4 +1,5 @@
-import { createContext } from 'react';
+import { createContext, ReactNode } from 'react';
+import { useState } from 'react';
 import { Product } from '../lib';
 
 // Define the type for the context value
@@ -15,3 +16,27 @@ const defaultCartValue: CartValue = {
 
 // Create the context
 export const CartContext = createContext<CartValue>(defaultCartValue);
+
+// Define the provider props types
+type CartProviderProps = {
+  children: ReactNode;
+};
+
+export function CartProvider({ children }: CartProviderProps) {
+  const [cartContent, setCartContent] = useState<Product[]>([]);
+
+  function addToCart(product: Product) {
+    setCartContent((prevCart) => [...prevCart, product]);
+  }
+
+  const cartContextValues = {
+    cart: cartContent,
+    addToCart,
+  };
+
+  return (
+    <CartContext.Provider value={cartContextValues}>
+      {children}
+    </CartContext.Provider>
+  );
+}
